voice-notes-starter/requirements.txt
streamlit>=1.36.0
openai>=1.40.0
python-dotenv>=1.0.1
pydub>=0.25.1
markdown>=3.6
jinja2>=3.1.4
reportlab>=4.2.0
pdfkit>=1.0.0
PyPDF2>=3.0.1
